silence system-originated experimental warning for PartitionMapping
